testing for example endpoints

+++ routes/example.js
@@ -165,6 +165,11 @@ router.put('/:id', (req, res, next) => {
 router.delete('/:id', (req, res, next) => {
   const { id } = req.params;  
   
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    const err = new Error('The `id` is not valid');
+    err.status = 400;
+    return next(err);
+  }
   //const /*userId*/ = req.user.id;
 
   TestExample_1.findOneAndRemove({_id:id,/*userId*/})
